(#98) Update Debian & Ubuntu factsets

+++ facts/get_facts.sh
@@ -92,9 +92,6 @@ case "${osfamily}" in
   ;;
 
 'Debian')
-  if [[ "xenial" =~ ${lsbdistcodename} ]]; then
-    lsbdistcodename='wily'
-  fi
   if [[ "serena" =~ ${lsbdistcodename} ]]; then
     lsbdistcodename='xenial'
   fi
@@ -107,10 +104,11 @@ case "${osfamily}" in
     dpkg --install /tmp/puppetlabs-release-pc1.deb
     apt-get update
     for puppet_agent_version in 1.2.2 1.4.2 1.5.3; do
-      apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*
-      output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
-      mkdir -p $(dirname ${output_file})
-      facter --show-legacy -p -j | tee ${output_file}
+      if apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*; then
+        output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
+        mkdir -p $(dirname ${output_file})
+        facter --show-legacy -p -j | tee ${output_file}
+      fi
     done
     apt-get -y remove --purge puppetlabs-release-pc1
   fi
@@ -119,10 +117,11 @@ case "${osfamily}" in
     dpkg --install /tmp/puppet5-release.deb
     apt-get update
     for puppet_agent_version in 5.0.1 5.1.0 5.3.7 5.4.0 5.5.3; do
-      apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*
-      output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
-      mkdir -p $(dirname ${output_file})
-      facter --show-legacy -p -j | tee ${output_file}
+      if apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*; then
+        output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
+        mkdir -p $(dirname ${output_file})
+        facter --show-legacy -p -j | tee ${output_file}
+      fi
     done
     apt-get -y remove --purge puppet5-release
   fi
@@ -131,10 +130,11 @@ case "${osfamily}" in
     dpkg --install /tmp/puppet6-release.deb
     apt-get update
     for puppet_agent_version in 6.2.0 6.4.2 6.6.0; do
-      apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*
-      output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
-      mkdir -p $(dirname ${output_file})
-      facter --show-legacy -p -j | tee ${output_file}
+      if apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*; then
+        output_file="/vagrant/$(facter --version | cut -d. -f1,2)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
+        mkdir -p $(dirname ${output_file})
+        facter --show-legacy -p -j | tee ${output_file}
+      fi
     done
     apt-get -y remove --purge puppet6-release
   fi
